+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## Unreleased
-- /
+
+- removed explicit loading of the BuildConfig plugin
+- fixed classpath being resolved too early
 
 ## [1.4.1] - 2025-10-25